backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
dj-rest-auth==7.2.0
Django==6.0.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-camel-case==1.4.2
drf-spectacular==0.29.0
drf-writable-nested==0.7.2
et_xmlfile==2.0.0
filelock==3.25.0
h11==0.16.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
openpyxl==3.1.2
outcome==1.3.0.post0
pillow==12.1.1
psutil==7.2.2
psycopg2-binary==2.9.10
pycparser==3.0
PySocks==1.7.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
selenium==4.41.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
sqlparse==0.5.5
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
